Refactor json plugin into epics.conversion.json
