+++ base/uk.ac.stfc.isis.ibex.experimentdetails/META-INF/MANIFEST.MF
@@ -8,11 +8,11 @@ Require-Bundle: org.eclipse.ui,
  uk.ac.stfc.isis.ibex.model;bundle-version="1.0.0",
  org.eclipse.core.runtime,
  uk.ac.stfc.isis.ibex.instrument;bundle-version="1.0.0",
- uk.ac.stfc.isis.ibex.json;bundle-version="1.0.0",
  org.eclipse.jface,
  uk.ac.stfc.isis.ibex.preferences,
  uk.ac.stfc.isis.ibex.databases;bundle-version="1.0.0",
- uk.ac.stfc.isis.ibex.epics.switching
+ uk.ac.stfc.isis.ibex.epics.switching,
+ com.google.gson
 Bundle-RequiredExecutionEnvironment: JavaSE-1.7
 Export-Package: uk.ac.stfc.isis.ibex.experimentdetails,
  uk.ac.stfc.isis.ibex.experimentdetails.database
