+++ base/uk.ac.stfc.isis.ibex.experimentdetails/src/uk/ac/stfc/isis/ibex/experimentdetails/internal/UserDetailsSerialiser.java
@@ -19,8 +19,8 @@
 
 package uk.ac.stfc.isis.ibex.experimentdetails.internal;
 
+import uk.ac.stfc.isis.ibex.epics.conversion.json.JsonSerialisingConverter;
 import uk.ac.stfc.isis.ibex.experimentdetails.UserDetails;
-import uk.ac.stfc.isis.ibex.json.JsonSerialisingConverter;
 
 public class UserDetailsSerialiser extends JsonSerialisingConverter<UserDetails[]> {
