+++ base/uk.ac.stfc.isis.ibex.instrument/META-INF/MANIFEST.MF
@@ -10,7 +10,7 @@ Require-Bundle: uk.ac.stfc.isis.ibex.epics;visibility:=reexport,
  uk.ac.stfc.isis.ibex.model;bundle-version="1.0.0",
  uk.ac.stfc.isis.ibex.logger,
  org.csstudio.utility.pvmanager,
- uk.ac.stfc.isis.ibex.json;bundle-version="1.0.0"
+ com.google.gson;bundle-version="2.1.0"
 Bundle-ActivationPolicy: lazy
 Export-Package: uk.ac.stfc.isis.ibex.instrument,
  uk.ac.stfc.isis.ibex.instrument.channels,
